validacion de usuario + ocultar cosas para No ADMIN

+++ src/main/java/net/daw/service/FacturaService.java
@@ -11,6 +11,7 @@ import java.util.ArrayList;
 import javax.servlet.http.HttpServletRequest;
 import net.daw.bean.FacturaBean;
 import net.daw.bean.ReplyBean;
+import net.daw.bean.UsuarioBean;
 import net.daw.connection.publicinterface.ConnectionInterface;
 import net.daw.constant.ConnectionConstants;
 import net.daw.dao.FacturaDao;
@@ -27,24 +28,37 @@ public class FacturaService {
         ob = oRequest.getParameter("ob");
     }
 
+    protected Boolean checkPermission(String strMethodName) {
+        UsuarioBean oUsuarioBean = (UsuarioBean) oRequest.getSession().getAttribute("user");
+        if (oUsuarioBean != null) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+
     public ReplyBean get() throws Exception {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            Integer id = Integer.parseInt(oRequest.getParameter("id"));
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            FacturaDao oFacturaDao = new FacturaDao(oConnection, ob);
-            FacturaBean oFacturaBean = oFacturaDao.get(id, 1);
-            Gson oGson = new Gson();
-            oReplyBean = new ReplyBean(200, oGson.toJson(oFacturaBean));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: get method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
-        }
+        if (this.checkPermission("get")) {
+            try {
+                Integer id = Integer.parseInt(oRequest.getParameter("id"));
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                FacturaDao oFacturaDao = new FacturaDao(oConnection, ob);
+                FacturaBean oFacturaBean = oFacturaDao.get(id, 1);
+                Gson oGson = new Gson();
+                oReplyBean = new ReplyBean(200, oGson.toJson(oFacturaBean));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: get method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
 
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
+        }
         return oReplyBean;
 
     }
@@ -53,17 +67,21 @@ public class FacturaService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            Integer id = Integer.parseInt(oRequest.getParameter("id"));
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            FacturaDao oFacturaDao = new FacturaDao(oConnection, ob);
-            int iRes = oFacturaDao.remove(id);
-            oReplyBean = new ReplyBean(200, Integer.toString(iRes));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: remove method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
+        if (this.checkPermission("remove")) {
+            try {
+                Integer id = Integer.parseInt(oRequest.getParameter("id"));
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                FacturaDao oFacturaDao = new FacturaDao(oConnection, ob);
+                int iRes = oFacturaDao.remove(id);
+                oReplyBean = new ReplyBean(200, Integer.toString(iRes));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: remove method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
         }
         return oReplyBean;
 
@@ -73,19 +91,23 @@ public class FacturaService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            FacturaDao oFacturaDao = new FacturaDao(oConnection, ob);
-            int registros = oFacturaDao.getcount();
-            Gson oGson = new Gson();
-            oReplyBean = new ReplyBean(200, oGson.toJson(registros));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: getcount method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
-        }
+        if (this.checkPermission("getcount")) {
+            try {
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                FacturaDao oFacturaDao = new FacturaDao(oConnection, ob);
+                int registros = oFacturaDao.getcount();
+                Gson oGson = new Gson();
+                oReplyBean = new ReplyBean(200, oGson.toJson(registros));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: getcount method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
 
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
+        }
         return oReplyBean;
 
     }
@@ -94,20 +116,24 @@ public class FacturaService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            String strJsonFromClient = oRequest.getParameter("json");
-            Gson oGson = new Gson();
-            FacturaBean oFacturaBean = new FacturaBean();
-            oFacturaBean = oGson.fromJson(strJsonFromClient, FacturaBean.class);
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            FacturaDao oFacturaDao = new FacturaDao(oConnection, ob);
-            oFacturaBean = oFacturaDao.create(oFacturaBean);
-            oReplyBean = new ReplyBean(200, oGson.toJson(oFacturaBean));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: create method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
+        if (this.checkPermission("create")) {
+            try {
+                String strJsonFromClient = oRequest.getParameter("json");
+                Gson oGson = new Gson();
+                FacturaBean oFacturaBean = new FacturaBean();
+                oFacturaBean = oGson.fromJson(strJsonFromClient, FacturaBean.class);
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                FacturaDao oFacturaDao = new FacturaDao(oConnection, ob);
+                oFacturaBean = oFacturaDao.create(oFacturaBean);
+                oReplyBean = new ReplyBean(200, oGson.toJson(oFacturaBean));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: create method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
         }
         return oReplyBean;
     }
@@ -117,20 +143,24 @@ public class FacturaService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            String strJsonFromClient = oRequest.getParameter("json");
-            Gson oGson = new Gson();
-            FacturaBean oFacturaBean = new FacturaBean();
-            oFacturaBean = oGson.fromJson(strJsonFromClient, FacturaBean.class);
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            FacturaDao oFacturaDao = new FacturaDao(oConnection, ob);
-            iRes = oFacturaDao.update(oFacturaBean);
-            oReplyBean = new ReplyBean(200, Integer.toString(iRes));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: update method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
+        if (this.checkPermission("update")) {
+            try {
+                String strJsonFromClient = oRequest.getParameter("json");
+                Gson oGson = new Gson();
+                FacturaBean oFacturaBean = new FacturaBean();
+                oFacturaBean = oGson.fromJson(strJsonFromClient, FacturaBean.class);
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                FacturaDao oFacturaDao = new FacturaDao(oConnection, ob);
+                iRes = oFacturaDao.update(oFacturaBean);
+                oReplyBean = new ReplyBean(200, Integer.toString(iRes));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: update method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
         }
         return oReplyBean;
     }
@@ -139,21 +169,25 @@ public class FacturaService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            Integer iRpp = Integer.parseInt(oRequest.getParameter("rpp"));
-            Integer iPage = Integer.parseInt(oRequest.getParameter("page"));
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            FacturaDao oFacturaDao = new FacturaDao(oConnection, ob);
-            ArrayList<FacturaBean> alFacturaBean = oFacturaDao.getpage(iRpp, iPage, 1);
-            Gson oGson = new Gson();
-            oReplyBean = new ReplyBean(200, oGson.toJson(alFacturaBean));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: getpage method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
-        }
+        if (this.checkPermission("getpage")) {
+            try {
+                Integer iRpp = Integer.parseInt(oRequest.getParameter("rpp"));
+                Integer iPage = Integer.parseInt(oRequest.getParameter("page"));
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                FacturaDao oFacturaDao = new FacturaDao(oConnection, ob);
+                ArrayList<FacturaBean> alFacturaBean = oFacturaDao.getpage(iRpp, iPage, 1);
+                Gson oGson = new Gson();
+                oReplyBean = new ReplyBean(200, oGson.toJson(alFacturaBean));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: getpage method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
 
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
+        }
         return oReplyBean;
 
     }
@@ -162,20 +196,24 @@ public class FacturaService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            Integer id = Integer.parseInt(oRequest.getParameter("id"));
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            FacturaDao oFacturaDao = new FacturaDao(oConnection, ob);
-            int registros = oFacturaDao.getcountFacturaUser(id);
-            Gson oGson = new Gson();
-            oReplyBean = new ReplyBean(200, oGson.toJson(registros));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: getcount method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
-        }
+        if (this.checkPermission("getcountFacturaUser")) {
+            try {
+                Integer id = Integer.parseInt(oRequest.getParameter("id"));
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                FacturaDao oFacturaDao = new FacturaDao(oConnection, ob);
+                int registros = oFacturaDao.getcountFacturaUser(id);
+                Gson oGson = new Gson();
+                oReplyBean = new ReplyBean(200, oGson.toJson(registros));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: getcount method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
 
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
+        }
         return oReplyBean;
 
     }
@@ -184,22 +222,26 @@ public class FacturaService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            Integer id_usuario = Integer.parseInt(oRequest.getParameter("id"));
-            Integer iRpp = Integer.parseInt(oRequest.getParameter("rpp"));
-            Integer iPage = Integer.parseInt(oRequest.getParameter("page"));
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            FacturaDao oFacturaDao = new FacturaDao(oConnection, ob);
-            ArrayList<FacturaBean> alLineaBean = oFacturaDao.getpageXusuario(iRpp, iPage, id_usuario, 1);
-            Gson oGson = new Gson();
-            oReplyBean = new ReplyBean(200, oGson.toJson(alLineaBean));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: getLineaFactura method: " + ob + " object" + ex.getMessage(), ex);
-        } finally {
-            oConnectionPool.disposeConnection();
-        }
+        if (this.checkPermission("getpageXusuario")) {
+            try {
+                Integer id_usuario = Integer.parseInt(oRequest.getParameter("id"));
+                Integer iRpp = Integer.parseInt(oRequest.getParameter("rpp"));
+                Integer iPage = Integer.parseInt(oRequest.getParameter("page"));
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                FacturaDao oFacturaDao = new FacturaDao(oConnection, ob);
+                ArrayList<FacturaBean> alLineaBean = oFacturaDao.getpageXusuario(iRpp, iPage, id_usuario, 1);
+                Gson oGson = new Gson();
+                oReplyBean = new ReplyBean(200, oGson.toJson(alLineaBean));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: getLineaFactura method: " + ob + " object" + ex.getMessage(), ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
 
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
+        }
         return oReplyBean;
 
     }
