+++ src/main/java/net/daw/service/LineaService.java
@@ -12,6 +12,7 @@ import java.util.ArrayList;
 import javax.servlet.http.HttpServletRequest;
 import net.daw.bean.LineaBean;
 import net.daw.bean.ReplyBean;
+import net.daw.bean.UsuarioBean;
 import net.daw.connection.publicinterface.ConnectionInterface;
 import net.daw.constant.ConnectionConstants;
 import net.daw.dao.LineaDao;
@@ -29,24 +30,37 @@ public class LineaService {
         ob = oRequest.getParameter("ob");
     }
 
+    protected Boolean checkPermission(String strMethodName) {
+        UsuarioBean oUsuarioBean = (UsuarioBean) oRequest.getSession().getAttribute("user");
+        if (oUsuarioBean != null) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+
     public ReplyBean get() throws Exception {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            Integer idfactura = Integer.parseInt(oRequest.getParameter("id"));
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            LineaDao oLineaDao = new LineaDao(oConnection, ob);
-            LineaBean oLineaBean = oLineaDao.get(idfactura, 1);
-            Gson oGson = (new GsonBuilder()).excludeFieldsWithoutExposeAnnotation().create();
-            oReplyBean = new ReplyBean(200, oGson.toJson(oLineaBean));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: get method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
-        }
+        if (this.checkPermission("get")) {
+            try {
+                Integer idfactura = Integer.parseInt(oRequest.getParameter("id"));
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                LineaDao oLineaDao = new LineaDao(oConnection, ob);
+                LineaBean oLineaBean = oLineaDao.get(idfactura, 1);
+                Gson oGson = (new GsonBuilder()).excludeFieldsWithoutExposeAnnotation().create();
+                oReplyBean = new ReplyBean(200, oGson.toJson(oLineaBean));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: get method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
 
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
+        }
         return oReplyBean;
 
     }
@@ -55,17 +69,21 @@ public class LineaService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            Integer id = Integer.parseInt(oRequest.getParameter("id"));
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            LineaDao oLineaDao = new LineaDao(oConnection, ob);
-            int iRes = oLineaDao.remove(id);
-            oReplyBean = new ReplyBean(200, Integer.toString(iRes));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: remove method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
+        if (this.checkPermission("remove")) {
+            try {
+                Integer id = Integer.parseInt(oRequest.getParameter("id"));
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                LineaDao oLineaDao = new LineaDao(oConnection, ob);
+                int iRes = oLineaDao.remove(id);
+                oReplyBean = new ReplyBean(200, Integer.toString(iRes));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: remove method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
         }
         return oReplyBean;
 
@@ -75,19 +93,23 @@ public class LineaService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            LineaDao oLineaDao = new LineaDao(oConnection, ob);
-            int registros = oLineaDao.getcount();
-            Gson oGson = new Gson();
-            oReplyBean = new ReplyBean(200, oGson.toJson(registros));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: getcount method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
-        }
+        if (this.checkPermission("getcount")) {
+            try {
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                LineaDao oLineaDao = new LineaDao(oConnection, ob);
+                int registros = oLineaDao.getcount();
+                Gson oGson = new Gson();
+                oReplyBean = new ReplyBean(200, oGson.toJson(registros));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: getcount method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
 
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
+        }
         return oReplyBean;
 
     }
@@ -96,20 +118,24 @@ public class LineaService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            String strJsonFromClient = oRequest.getParameter("json");
-            Gson oGson = new Gson();
-            LineaBean oLineaBean = new LineaBean();
-            oLineaBean = oGson.fromJson(strJsonFromClient, LineaBean.class);
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            LineaDao oLineaDao = new LineaDao(oConnection, ob);
-            oLineaBean = oLineaDao.create(oLineaBean);
-            oReplyBean = new ReplyBean(200, oGson.toJson(oLineaBean));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: create method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
+        if (this.checkPermission("create")) {
+            try {
+                String strJsonFromClient = oRequest.getParameter("json");
+                Gson oGson = new Gson();
+                LineaBean oLineaBean = new LineaBean();
+                oLineaBean = oGson.fromJson(strJsonFromClient, LineaBean.class);
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                LineaDao oLineaDao = new LineaDao(oConnection, ob);
+                oLineaBean = oLineaDao.create(oLineaBean);
+                oReplyBean = new ReplyBean(200, oGson.toJson(oLineaBean));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: create method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
         }
         return oReplyBean;
     }
@@ -119,6 +145,7 @@ public class LineaService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
+        if (this.checkPermission("update")) {
             try {
                 String strJsonFromClient = oRequest.getParameter("json");
                 Gson oGson = (new GsonBuilder()).excludeFieldsWithoutExposeAnnotation().create();
@@ -135,28 +162,35 @@ public class LineaService {
             } finally {
                 oConnectionPool.disposeConnection();
             }
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
+        }
         return oReplyBean;
-}
+    }
 
     public ReplyBean getpage() throws Exception {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            Integer iRpp = Integer.parseInt(oRequest.getParameter("rpp"));
-            Integer iPage = Integer.parseInt(oRequest.getParameter("page"));
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            LineaDao oLineaDao = new LineaDao(oConnection, ob);
-            ArrayList<LineaBean> alLineaBean = oLineaDao.getpage(iRpp, iPage);
-            Gson oGson = new Gson();
-            oReplyBean = new ReplyBean(200, oGson.toJson(alLineaBean));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: getpage method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
-        }
+        if (this.checkPermission("getpage")) {
+            try {
+                Integer iRpp = Integer.parseInt(oRequest.getParameter("rpp"));
+                Integer iPage = Integer.parseInt(oRequest.getParameter("page"));
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                LineaDao oLineaDao = new LineaDao(oConnection, ob);
+                ArrayList<LineaBean> alLineaBean = oLineaDao.getpage(iRpp, iPage);
+                Gson oGson = new Gson();
+                oReplyBean = new ReplyBean(200, oGson.toJson(alLineaBean));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: getpage method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
 
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
+        }
         return oReplyBean;
 
     }
@@ -165,22 +199,26 @@ public class LineaService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            Integer id_factura = Integer.parseInt(oRequest.getParameter("id"));
-            Integer iRpp = Integer.parseInt(oRequest.getParameter("rpp"));
-            Integer iPage = Integer.parseInt(oRequest.getParameter("page"));
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            LineaDao oLineaDao = new LineaDao(oConnection, ob);
-            ArrayList<LineaBean> alLineaBean = oLineaDao.getLineaFactura(iRpp, iPage, id_factura, 1);
-            Gson oGson = new Gson();
-            oReplyBean = new ReplyBean(200, oGson.toJson(alLineaBean));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: getLineaFactura method: " + ob + " object" + ex.getMessage(), ex);
-        } finally {
-            oConnectionPool.disposeConnection();
-        }
+        if (this.checkPermission("getLineaFactura")) {
+            try {
+                Integer id_factura = Integer.parseInt(oRequest.getParameter("id"));
+                Integer iRpp = Integer.parseInt(oRequest.getParameter("rpp"));
+                Integer iPage = Integer.parseInt(oRequest.getParameter("page"));
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                LineaDao oLineaDao = new LineaDao(oConnection, ob);
+                ArrayList<LineaBean> alLineaBean = oLineaDao.getLineaFactura(iRpp, iPage, id_factura, 1);
+                Gson oGson = new Gson();
+                oReplyBean = new ReplyBean(200, oGson.toJson(alLineaBean));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: getLineaFactura method: " + ob + " object" + ex.getMessage(), ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
 
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
+        }
         return oReplyBean;
 
     }
@@ -189,20 +227,24 @@ public class LineaService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            Integer id_factura = Integer.parseInt(oRequest.getParameter("id"));
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            LineaDao oLineaDao = new LineaDao(oConnection, ob);
-            int registros = oLineaDao.getcountxlinea(id_factura);
-            Gson oGson = new Gson();
-            oReplyBean = new ReplyBean(200, oGson.toJson(registros));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: getcount method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
-        }
+        if (this.checkPermission("getcountxlinea")) {
+            try {
+                Integer id_factura = Integer.parseInt(oRequest.getParameter("id"));
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                LineaDao oLineaDao = new LineaDao(oConnection, ob);
+                int registros = oLineaDao.getcountxlinea(id_factura);
+                Gson oGson = new Gson();
+                oReplyBean = new ReplyBean(200, oGson.toJson(registros));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: getcount method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
 
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
+        }
         return oReplyBean;
 
     }
